% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRM_Ion_Pair_Finder.R
\name{MRM_Ion_Pair_Finder}
\alias{MRM_Ion_Pair_Finder}
\title{Title MRM_Ion_Pair_Finder}
\usage{
MRM_Ion_Pair_Finder(
  file_MS1,
  filepath_MS2,
  tol_mz,
  tol_tr,
  diff_MS2MS1,
  ms2_intensity,
  resultpath,
  OnlyKeepChargeEqual1 = TRUE,
  cores = 4
)
}
\arguments{
\item{file_MS1}{MS1 peak detection result save in .csv filetype, the first column is m/z named 'mz',
the second column is retention time(s) named 'tr',
intensity of samples is located begin the third column.}

\item{filepath_MS2}{The folder path which have mgf files.}

\item{tol_mz}{The tolerence of m/z between MS1 peak detection result and mgf files. 0.01 is suitable for Q-TOF.}

\item{tol_tr}{The tolerence of retention time between MS1 peak detection result and mgf files.}

\item{diff_MS2MS1}{The smallest difference between product ion and precusor ion.}

\item{ms2_intensity}{The smallest intensity of product ion.}

\item{resultpath}{A csv file named "MRM transitions list.csv" will saved in the path.}

\item{OnlyKeepChargeEqual1}{T or F}
}
\value{
data_ms1ms2_final
}
\description{
Title MRM_Ion_Pair_Finder
}
